refactored DocumentTransformer, FacetedTransformer, EventQueryTransformer  Changes to be committed: 	modified:   warehouse/query-core/src/main/java/datawave/query/transformer/DocumentTransformer.java 	renamed:    warehouse/query-core/src/main/java/datawave/query/transformer/DocumentTransformer.java -> warehouse/query-core/src/main/java/datawave/query/transformer/DocumentTransformerSupport.java 	modified:   warehouse/query-core/src/main/java/datawave/query/transformer/EventQueryTransformer.java 	new file:   warehouse/query-core/src/main/java/datawave/query/transformer/EventQueryTransformerSupport.java 	modified:   warehouse/query-core/src/main/java/datawave/query/transformer/FacetedTransformer.java
